+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 CHANGELOG
 =========
 
-
-UNRELEASED
-----------
+unreleased
+-----------
+- Add script tag integration for cards only
+- Add support for custom translations
 - Use npm scoped version of browser-detection
 - Clean up payment option error messages
 
@@ -15,6 +16,7 @@ UNRELEASED
 - Fix width errors where Drop-in was not aligned with other elements on merchant page
 - Add ability to style PayPal button
 - Fail early if PayPal creation errors
+- Allow card overrides with Hosted Fields
 - Upgrade braintree-web to v3.17.0
 - Upgrade checkout.js to 4.0.78